requirements.txt
Flask==2.0.2

Flask-SQLAlchemy==2.5.1

PyJWT==2.3.0

SQLAlchemy==1.4.26